Reinstate explicit unboxing in the few places it's needed.
git-svn-id: https://svn.code.sf.net/p/vassalengine/svn/VASSAL-src/trunk@9400 67b53d14-2c14-4ace-a08f-0dab2b34000c

+++ test/VASSAL/tools/concurrent/listener/EventAccumulatorTest.java
@@ -36,7 +36,7 @@ public class EventAccumulatorTest {
     int i = 0;
     for (Pair<Object,Integer> e : ea.events()) {
       assertSame(this, e.first);
-      assertEquals(i, e.second);
+      assertEquals(i, e.second.intValue());
       ++i;
     }
   }
